Added String based property
Do not keep SettingBase reference
Refactored  ViewModels and Views

+++ Actions/SetFloatSettingProperty.cs
@@ -15,7 +15,7 @@ namespace MBOptionScreen.Actions
         {
             Value = value;
             SettingProperty = settingProperty;
-            _originalValue = SettingProperty.IntValue;
+            _originalValue = SettingProperty.FloatValue;
         }
 
         public void DoAction() => SettingProperty.FloatValue = (float)Value;
